Move fold graph into library
